Make some notes
Trying to understand WAM a bit. There are probably ways
to avoid the manual stacks and indirection and leave the
Go GC to do the work. I think golog figured that out.

But understanding the WAM seems like a reasonable path to
enlightenment, so will attempt o implement in order.